requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
pymongo[srv]==4.10.1
flask-restx==1.3.0
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
werkzeug==3.0.4
